+++ appyx-components/stable/spotlight/common/src/commonMain/kotlin/com/bumble/appyx/components/spotlight/SpotlightModel.kt
@@ -54,6 +54,9 @@ class SpotlightModel<InteractionTarget : Any>(
             activeIndex = initialActiveIndex
         )
 
+    val currentState: State<InteractionTarget>
+        get() = output.value.currentTargetState
+
     override fun State<InteractionTarget>.removeDestroyedElement(
         element: Element<InteractionTarget>
     ): State<InteractionTarget> {