+++ appyx-components/stable/spotlight/common/src/commonMain/kotlin/com/bumble/appyx/components/spotlight/ui/sliderrotation/SpotlightSliderRotation.kt
@@ -20,15 +20,16 @@ import com.bumble.appyx.transitionmodel.BaseVisualisation
 
 class SpotlightSliderRotation<InteractionTarget : Any>(
     uiContext: UiContext,
+    initialState: State<InteractionTarget>,
     @Suppress("UnusedPrivateMember")
     private val orientation: Orientation = Orientation.Horizontal, // TODO support RTL
 ) : BaseVisualisation<InteractionTarget, State<InteractionTarget>, MutableUiState, TargetUiState>(
     uiContext = uiContext
 ) {
     private val scrollX = GenericFloatProperty(
-        uiContext.coroutineScope,
-        Target(0f)
-    ) // TODO sync this with the model's initial value rather than assuming 0
+        coroutineScope = uiContext.coroutineScope,
+        target = Target(initialState.activeIndex),
+    )
     override val viewpointDimensions: List<Pair<(State<InteractionTarget>) -> Float, GenericFloatProperty>> =
         listOf(
             { state: State<InteractionTarget> -> state.activeIndex } to scrollX