+++ appyx-interactions/common/src/commonMain/kotlin/com/bumble/appyx/transitionmodel/BaseVisualisation.kt
@@ -97,8 +97,9 @@ abstract class BaseVisualisation<InteractionTarget : Any, ModelState, MutableUiS
 
         cleanUpCacheForDestroyedElements(matchedTargetUiStates)
 
+        val isRecreating = mutableUiStateCache.isEmpty()
         coroutineScope.launch {
-            updateViewpoint(update)
+            updateViewpoint(update, isRecreating)
         }
 
 
@@ -118,7 +119,7 @@ abstract class BaseVisualisation<InteractionTarget : Any, ModelState, MutableUiS
                     ObserveElementAnimationChanges(mutableUiState, t1)
                     ManageAnimations(mutableUiState, t1, update)
                 },
-                motionProperties = mutableUiState.motionProperties,
+                motionProperties = mutableUiState.motionProperties + viewpointDimensions.map { it.second },
                 modifier = mutableUiState.modifier,
                 progress = MutableStateFlow(1f),
             )
@@ -193,17 +194,24 @@ abstract class BaseVisualisation<InteractionTarget : Any, ModelState, MutableUiS
         }
     }
 
-    private suspend fun updateViewpoint(update: Update<ModelState>) {
+    private suspend fun updateViewpoint(
+        update: Update<ModelState>,
+        isRecreating: Boolean
+    ) {
         viewpointDimensions.forEach { (fieldOfState, viewpointDimension) ->
             val targetValue = fieldOfState(update.currentTargetState)
-            viewpointDimension.animateTo(
-                targetValue,
-                spring(
-                    stiffness = currentSpringSpec.stiffness,
-                    dampingRatio = currentSpringSpec.dampingRatio
-                )
-            ) {
-                AppyxLogger.d(TAG, "Viewpoint animateTo (Update) – $targetValue")
+            if (isRecreating) {
+                viewpointDimension.snapTo(targetValue)
+            } else {
+                viewpointDimension.animateTo(
+                    targetValue,
+                    spring(
+                        stiffness = currentSpringSpec.stiffness,
+                        dampingRatio = currentSpringSpec.dampingRatio
+                    )
+                ) {
+                    AppyxLogger.d(TAG, "Viewpoint animateTo (Update) – $targetValue")
+                }
             }
         }
     }
@@ -245,7 +253,7 @@ abstract class BaseVisualisation<InteractionTarget : Any, ModelState, MutableUiS
                     Box(modifier = mutableUiState.visibilityModifier)
                     InterpolateUiState(segmentProgress, mutableUiState, t0, t1, initialProgress)
                 },
-                motionProperties = mutableUiState.motionProperties,
+                motionProperties = mutableUiState.motionProperties + viewpointDimensions.map { it.second },
                 modifier = mutableUiState.modifier,
                 progress = segmentProgress,
             )