+++ demos/appyx-interactions/android/src/main/kotlin/com/bumble/appyx/interactions/bottomnav/MainNavItem.kt
@@ -2,19 +2,18 @@ package com.bumble.appyx.interactions.bottomnav
 
 import androidx.compose.material.icons.Icons.Filled
 import androidx.compose.material.icons.Icons.Outlined
+import androidx.compose.material.icons.filled.Cake
 import androidx.compose.material.icons.filled.Extension
-import androidx.compose.material.icons.filled.Style
 import androidx.compose.material.icons.filled.ViewCarousel
+import androidx.compose.material.icons.outlined.Cake
 import androidx.compose.material.icons.outlined.Extension
 import androidx.compose.material.icons.outlined.GridView
-import androidx.compose.material.icons.outlined.Style
 import androidx.compose.material.icons.outlined.ViewCarousel
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.rotate
 import com.bumble.appyx.components.experimental.cards.android.DatingCards
 import com.bumble.appyx.components.experimental.puzzle15.android.Puzzle15
 import com.bumble.appyx.components.internal.testdrive.android.TestDriveExperiment
-import com.bumble.appyx.components.spotlight.ui.sliderrotation.SpotlightSliderRotation
 import com.bumble.appyx.interactions.sample.SpotlightExperiment
 import com.bumble.appyx.navigation.node.node
 import com.bumble.appyx.utils.material3.AppyxNavItem
@@ -30,9 +29,9 @@ enum class MainNavItem : Parcelable {
             when (navBarItem) {
                 CARDS -> AppyxNavItem(
                     text = "Cards",
-                    unselectedIcon = Outlined.Style,
-                    selectedIcon = Filled.Style,
-                    iconModifier = Modifier.rotate(180f),
+                    unselectedIcon = Outlined.Cake,
+                    selectedIcon = Filled.Cake,
+                    modifier = Modifier.rotate(180f),
                     node = { node(it) { DatingCards() } }
                 )
 
@@ -40,7 +39,7 @@ enum class MainNavItem : Parcelable {
                     text = "Spotlight",
                     unselectedIcon = Outlined.ViewCarousel,
                     selectedIcon = Filled.ViewCarousel,
-                    node = { node(it) { SpotlightExperiment { SpotlightSliderRotation(it) } } }
+                    node = { node(it) { SpotlightExperiment() } }
                 )
 
                 TEST_DRIVE -> AppyxNavItem(