+++ demos/appyx-navigation/common/src/androidMain/kotlin/com/bumble/appyx/navigation/platform/PlatformName.kt
@@ -1,0 +1,3 @@
+package com.bumble.appyx.navigation.platform
+
+actual fun getPlatformName(): String = "Android"