+++ demos/appyx-navigation/common/src/commonMain/kotlin/com/bumble/appyx/navigation/node/checkout/CheckoutFormField.kt
@@ -1,0 +1,5 @@
+package com.bumble.appyx.navigation.node.checkout
+
+interface CheckoutFormField {
+    val value: String
+}