+++ demos/appyx-navigation/common/src/commonMain/kotlin/com/bumble/appyx/navigation/ui/Theme.kt
@@ -2,35 +2,21 @@ package com.bumble.appyx.navigation.ui
 
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Typography
 import androidx.compose.material3.darkColorScheme
 import androidx.compose.material3.lightColorScheme
 import androidx.compose.runtime.Composable
 
 
-private val DarkColorPalette = darkColorScheme(
-    primary = appyx_yellow1,
-    secondary = appyx_yellow2,
-    background = appyx_dark,
-    surface = appyx_dark,
-    onPrimary = appyx_bright,
-    onSecondary = appyx_bright,
-    onBackground = appyx_bright,
-    onSurface = appyx_bright,
-)
-
-private val LightColorPalette = lightColorScheme(
-    primary = appyx_yellow1,
-    secondary = appyx_yellow2,
-    background = appyx_bright,
-    surface = appyx_bright,
-    onPrimary = appyx_dark,
-    onSecondary = appyx_dark,
-    onBackground = appyx_dark,
-    onSurface = appyx_dark,
-)
+private val DarkColorPalette = darkColorScheme()
+private val LightColorPalette = lightColorScheme()
 
 @Composable
-fun AppyxSampleAppTheme(darkTheme: Boolean = isSystemInDarkTheme(), content: @Composable () -> Unit) {
+fun AppyxSampleAppTheme(
+    darkTheme: Boolean = isSystemInDarkTheme(),
+    themeTypography: Typography = typography,
+    content: @Composable () -> Unit
+) {
     val colors = if (darkTheme) {
         DarkColorPalette
     } else {
@@ -39,7 +25,7 @@ fun AppyxSampleAppTheme(darkTheme: Boolean = isSystemInDarkTheme(), content: @Co
 
     MaterialTheme(
         colorScheme = colors,
-        typography = typography,
+        typography = themeTypography,
         shapes = shapes,
         content = content
     )