+++ demos/appyx-navigation/ios/src/iosMain/kotlin/main.ios.kt
@@ -1,62 +1,47 @@
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.WindowInsets
-import androidx.compose.foundation.layout.asPaddingValues
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.statusBars
-import androidx.compose.foundation.layout.windowInsetsPadding
-import androidx.compose.material.Icon
-import androidx.compose.material.IconButton
 import androidx.compose.material.Scaffold
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.ArrowBack
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.ComposeUIViewController
-import androidx.compose.ui.zIndex
 import com.bumble.appyx.navigation.integration.IosNodeHost
 import com.bumble.appyx.navigation.integration.MainIntegrationPoint
-import com.bumble.appyx.navigation.node.container.MainNavNode
+import com.bumble.appyx.navigation.navigator.LocalNavigator
+import com.bumble.appyx.navigation.navigator.Navigator
+import com.bumble.appyx.navigation.node.root.RootNode
 import com.bumble.appyx.navigation.ui.AppyxSampleAppTheme
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.channels.Channel
-import kotlinx.coroutines.flow.receiveAsFlow
-import kotlinx.coroutines.launch
-
-val backEvents: Channel<Unit> = Channel()
+import kotlinx.coroutines.flow.flowOf
+import platform.Foundation.NSURL
 
 private val integrationPoint = MainIntegrationPoint()
+private val navigator = Navigator()
 
 @Suppress("FunctionNaming")
 fun MainViewController() = ComposeUIViewController {
-
     AppyxSampleAppTheme {
-        val coroutineScope = rememberCoroutineScope()
         Scaffold(
             modifier = Modifier
                 .background(Color.Black)
+//                .padding(top = 60.dp)
         ) {
             Box {
-                BackButton(
-                    coroutineScope = coroutineScope,
-                    modifier = Modifier.padding(
-                        top = WindowInsets.statusBars.asPaddingValues().calculateTopPadding()
-                    )
-                )
 
                 Box(modifier = Modifier.fillMaxSize()) {
-                    IosNodeHost(
-                        modifier = Modifier,
-                        onBackPressedEvents = backEvents.receiveAsFlow(),
-                        integrationPoint = remember { integrationPoint }
-                    ) { buildContext ->
-                        MainNavNode(
-                            buildContext = buildContext,
-                        )
+                    CompositionLocalProvider(LocalNavigator provides navigator) {
+                        IosNodeHost(
+                            modifier = Modifier,
+                            onBackPressedEvents = flowOf(),
+                            integrationPoint = integrationPoint
+                        ) { buildContext ->
+                            RootNode(
+                                buildContext = buildContext,
+                                plugins = listOf(navigator),
+                            )
+                        }
                     }
                 }
             }
@@ -66,23 +51,15 @@ fun MainViewController() = ComposeUIViewController {
     integrationPoint.setViewController(uiViewController)
 }
 
-@Composable
-private fun BackButton(
-    coroutineScope: CoroutineScope,
-    modifier: Modifier = Modifier,
-) {
-    IconButton(
-        onClick = {
-            coroutineScope.launch {
-                backEvents.send(Unit)
-            }
-        },
-        modifier = modifier.zIndex(99f)
-    ) {
-        Icon(
-            imageVector = Icons.Default.ArrowBack,
-            tint = Color.White,
-            contentDescription = "Go Back",
-        )
+@Suppress("FunctionNaming")
+fun handleDeepLinks(url: NSURL) {
+    if (url.scheme == "appyx") {
+        when (url.host) {
+            // xcrun simctl openurl booted 'appyx://randomcake'
+            ("randomcake") -> navigator.goToARandomCakeWithDummyUser()
+
+            // xcrun simctl openurl booted 'appyx://randomcake-wait'
+            ("randomcake-wait") -> navigator.goToARandomCake()
+        }
     }
 }